Move util to shared folder
