+++ src/modules/parking/components/sessionsList/ParkingSessionHistoryList.tsx
@@ -5,7 +5,6 @@ import {Box} from '@/components/ui/containers/Box'
 import {Gutter} from '@/components/ui/layout/Gutter'
 import {Phrase} from '@/components/ui/text/Phrase'
 import {useInfiniteScroller} from '@/hooks/useInfiniteScroller'
-import {getCurrentPage} from '@/modules/construction-work/components/projects/utils/getCurrentPage'
 import {ParkingSessionListRenderItem} from '@/modules/parking/components/sessionsList/ParkingSessionListRenderItem'
 import {useCurrentParkingPermit} from '@/modules/parking/hooks/useCurrentParkingPermit'
 import {
@@ -23,6 +22,7 @@ import {
   groupParkingSessionsByDate,
 } from '@/modules/parking/utils/groupParkingSessionsByDate'
 import {layoutStyles} from '@/styles/layoutStyles'
+import {getCurrentPage} from '@/utils/pagination/getCurrentPage'
 
 type ParkingHistorySessionOrDummy =
   | (ParkingHistorySession & {dummy?: never})
@@ -42,7 +42,6 @@ export const ParkingSessionHistoryList = ({
   sortAscending = false,
 }: Props) => {
   const currentPermit = useCurrentParkingPermit()
-
   const [viewableItemIndex, setViewableItemIndex] = useState(1)
   const page = getCurrentPage(viewableItemIndex, 1, pageSize)
